create htmlview class and inject it to game
